Updated pins and comments
Changed all the comments to better represent what the PCB has wired. Some of these include pin 9 which was previously unused. Additionally, we removed the commented pin 12 which used to be the input for L293D on one side. This reflects the wiring change we made to use pin 10 for both L293D inputs